+++ packages/wenyou_api/lib/src/model/profile_cover_response_dto.g.dart
@@ -15,6 +15,8 @@ class _$ProfileCoverResponseDto extends ProfileCoverResponseDto {
   final num? width;
   @override
   final num? height;
+  @override
+  final ProfileCoverVariantResponseDto? mobile;
 
   factory _$ProfileCoverResponseDto([
     void Function(ProfileCoverResponseDtoBuilder)? updates,
@@ -25,6 +27,7 @@ class _$ProfileCoverResponseDto extends ProfileCoverResponseDto {
     this.mediumUrl,
     this.width,
     this.height,
+    this.mobile,
   }) : super._();
   @override
   ProfileCoverResponseDto rebuild(
@@ -42,7 +45,8 @@ class _$ProfileCoverResponseDto extends ProfileCoverResponseDto {
         url == other.url &&
         mediumUrl == other.mediumUrl &&
         width == other.width &&
-        height == other.height;
+        height == other.height &&
+        mobile == other.mobile;
   }
 
   @override
@@ -52,6 +56,7 @@ class _$ProfileCoverResponseDto extends ProfileCoverResponseDto {
     _$hash = $jc(_$hash, mediumUrl.hashCode);
     _$hash = $jc(_$hash, width.hashCode);
     _$hash = $jc(_$hash, height.hashCode);
+    _$hash = $jc(_$hash, mobile.hashCode);
     _$hash = $jf(_$hash);
     return _$hash;
   }
@@ -62,7 +67,8 @@ class _$ProfileCoverResponseDto extends ProfileCoverResponseDto {
           ..add('url', url)
           ..add('mediumUrl', mediumUrl)
           ..add('width', width)
-          ..add('height', height))
+          ..add('height', height)
+          ..add('mobile', mobile))
         .toString();
   }
 }
@@ -88,6 +94,12 @@ class ProfileCoverResponseDtoBuilder
   num? get height => _$this._height;
   set height(num? height) => _$this._height = height;
 
+  ProfileCoverVariantResponseDtoBuilder? _mobile;
+  ProfileCoverVariantResponseDtoBuilder get mobile =>
+      _$this._mobile ??= ProfileCoverVariantResponseDtoBuilder();
+  set mobile(ProfileCoverVariantResponseDtoBuilder? mobile) =>
+      _$this._mobile = mobile;
+
   ProfileCoverResponseDtoBuilder() {
     ProfileCoverResponseDto._defaults(this);
   }
@@ -99,6 +111,7 @@ class ProfileCoverResponseDtoBuilder
       _mediumUrl = $v.mediumUrl;
       _width = $v.width;
       _height = $v.height;
+      _mobile = $v.mobile?.toBuilder();
       _$v = null;
     }
     return this;
@@ -118,18 +131,35 @@ class ProfileCoverResponseDtoBuilder
   ProfileCoverResponseDto build() => _build();
 
   _$ProfileCoverResponseDto _build() {
-    final _$result =
-        _$v ??
-        _$ProfileCoverResponseDto._(
-          url: BuiltValueNullFieldError.checkNotNull(
-            url,
-            r'ProfileCoverResponseDto',
-            'url',
-          ),
-          mediumUrl: mediumUrl,
-          width: width,
-          height: height,
+    _$ProfileCoverResponseDto _$result;
+    try {
+      _$result =
+          _$v ??
+          _$ProfileCoverResponseDto._(
+            url: BuiltValueNullFieldError.checkNotNull(
+              url,
+              r'ProfileCoverResponseDto',
+              'url',
+            ),
+            mediumUrl: mediumUrl,
+            width: width,
+            height: height,
+            mobile: _mobile?.build(),
+          );
+    } catch (_) {
+      late String _$failedField;
+      try {
+        _$failedField = 'mobile';
+        _mobile?.build();
+      } catch (e) {
+        throw BuiltValueNestedFieldError(
+          r'ProfileCoverResponseDto',
+          _$failedField,
+          e.toString(),
         );
+      }
+      rethrow;
+    }
     replace(_$result);
     return _$result;
   }