+++ packages/wenyou_api/lib/src/serializers.dart
@@ -293,6 +293,7 @@ import 'package:wenyou_api/src/model/posts_update200_response.dart';
 import 'package:wenyou_api/src/model/posts_upsert_body200_response.dart';
 import 'package:wenyou_api/src/model/private_user_response_dto.dart';
 import 'package:wenyou_api/src/model/profile_cover_response_dto.dart';
+import 'package:wenyou_api/src/model/profile_cover_variant_response_dto.dart';
 import 'package:wenyou_api/src/model/progression_response_dto.dart';
 import 'package:wenyou_api/src/model/public_user_response_dto.dart';
 import 'package:wenyou_api/src/model/recent_reply_dice_response_dto.dart';
@@ -745,6 +746,7 @@ part 'serializers.g.dart';
   PostsUpsertBody200Response,
   PrivateUserResponseDto,
   ProfileCoverResponseDto,
+  ProfileCoverVariantResponseDto,
   ProgressionResponseDto,
   PublicUserResponseDto,
   RecentReplyDiceResponseDto,